requirements.txt
aiogram==2.23.1
firebase-admin==6.5.0
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
